Binder ready for energy model
